requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
tensorflow==2.15.0
autokeras==1.1.0
open-clip-torch==2.24.0
torch==2.1.1
torchvision==0.16.1
requests==2.31.0
